Removed and Added code for Domain objects

+++ src/main/java/org/jax/mgi/mgd/api/controller/VocabController.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 import javax.inject.Inject;
 
-import org.jax.mgi.mgd.api.model.voc.entities.Vocabulary;
+import org.jax.mgi.mgd.api.domain.VocabularyDomain;
 import org.jax.mgi.mgd.api.rest.interfaces.VocabRESTInterface;
 import org.jax.mgi.mgd.api.service.VocabService;
 import org.jax.mgi.mgd.api.util.SearchResults;
@@ -15,27 +15,36 @@ public class VocabController extends BaseController implements VocabRESTInterfac
 	private VocabService vocabService;
 
 	@Override
-	public Vocabulary create(String api_access_token, Vocabulary object) {
-		return vocabService.create(object);
+	public VocabularyDomain create(String api_access_token, VocabularyDomain object) {
+		if(authenticate(api_access_token)) {
+			return vocabService.create(object);
+		}
+		return null;
 	}
 
 	@Override
-	public Vocabulary get(Integer key) {
+	public VocabularyDomain get(Integer key) {
 		return vocabService.get(key);
 	}
 
 	@Override
-	public Vocabulary update(String api_access_token, Vocabulary object) {
-		return vocabService.update(object);
+	public VocabularyDomain update(String api_access_token, VocabularyDomain object) {
+		if(authenticate(api_access_token)) {
+			return vocabService.update(object);
+		}
+		return null;
 	}
 
 	@Override
-	public Vocabulary delete(String api_access_token, Integer key) {
-		return vocabService.delete(key);
+	public VocabularyDomain delete(String api_access_token, Integer key) {
+		if(authenticate(api_access_token)) {
+			return vocabService.delete(key);
+		}
+		return null;
 	}
 
 	@Override
-	public SearchResults<Vocabulary> search(Map<String, Object> postParams) {
+	public SearchResults<VocabularyDomain> search(Map<String, Object> postParams) {
 		return vocabService.search(postParams);
 	}
